@@ -10,24 +10,24 @@
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-node": "^4.0.1",
-		"@sveltejs/kit": "^2.5.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.2",
-		"@types/cookie": "^0.5.1",
-		"@types/node": "^20.5.9",
-		"@types/uuid": "^9.0.3",
-		"svelte": "^5.0.0-next.57",
-		"svelte-check": "^3.6.4",
-		"svelte-preprocess": "^5.1.3",
+		"@sveltejs/adapter-node": "^5.0.1",
+		"@sveltejs/kit": "^2.5.7",
+		"@sveltejs/vite-plugin-svelte": "^3.1.0",
+		"@types/cookie": "^0.6.0",
+		"@types/node": "^20.12.11",
+		"@types/uuid": "^9.0.8",
+		"svelte": "^5.0.0-next.126",
+		"svelte-check": "^3.7.1",
+		"svelte-preprocess": "^5.1.4",
 		"tsconfig": "^7.0.0",
 		"tslib": "^2.6.2",
-		"typescript": "^5.3.3",
-		"vite": "^5.1.3"
+		"typescript": "^5.4.5",
+		"vite": "^5.2.11"
 	},
 	"dependencies": {
-		"cm-chessboard": "^8.5.0",
-		"nanoid": "^5.0.6",
-		"uuid": "^9.0.0",
+		"cm-chessboard": "^8.7.4",
+		"nanoid": "^5.0.7",
+		"uuid": "^9.0.1",
 		"dotenv": "^16.4.5"
 	},
 	"type": "module"
